refactor: extract tui breakdowns and split daemon server
