+++ internal/daemon/server.go
@@ -19,11 +19,9 @@ import (
 
 	"github.com/janekbaraniewski/openusage/internal/config"
 	"github.com/janekbaraniewski/openusage/internal/core"
-	"github.com/janekbaraniewski/openusage/internal/integrations"
 	"github.com/janekbaraniewski/openusage/internal/providers"
 	"github.com/janekbaraniewski/openusage/internal/providers/shared"
 	"github.com/janekbaraniewski/openusage/internal/telemetry"
-	"github.com/janekbaraniewski/openusage/internal/version"
 )
 
 type Service struct {
@@ -205,129 +203,6 @@ func (s *Service) flushBacklog(ctx context.Context, retryReqs []telemetry.Ingest
 	s.spoolMu.Unlock()
 
 	return flush, enqueued, append(warnings, flushWarnings...)
-}
-
-// --- Logging ---
-
-func (s *Service) infof(event, format string, args ...any) {
-	if s == nil || !s.cfg.Verbose {
-		return
-	}
-	if strings.TrimSpace(format) == "" {
-		log.Printf("daemon level=info event=%s", event)
-		return
-	}
-	log.Printf("daemon level=info event=%s "+format, append([]any{event}, args...)...)
-}
-
-func (s *Service) warnf(event, format string, args ...any) {
-	if s == nil || !s.cfg.Verbose {
-		return
-	}
-	if strings.TrimSpace(format) == "" {
-		log.Printf("daemon level=warn event=%s", event)
-		return
-	}
-	log.Printf("daemon level=warn event=%s "+format, append([]any{event}, args...)...)
-}
-
-func (s *Service) shouldLog(key string, interval time.Duration) bool {
-	if s == nil {
-		return false
-	}
-	return s.logThrottle.Allow(key, interval, time.Now())
-}
-
-// --- Read-model cache ---
-
-func (s *Service) computeReadModel(
-	ctx context.Context,
-	req ReadModelRequest,
-) (map[string]core.UsageSnapshot, error) {
-	start := time.Now()
-	templates := ReadModelTemplatesFromRequest(req, DisabledAccountsFromConfig())
-	if len(templates) == 0 {
-		return map[string]core.UsageSnapshot{}, nil
-	}
-	tw := normalizeReadModelTimeWindow(req.TimeWindow)
-	result, err := telemetry.ApplyCanonicalTelemetryViewWithOptions(ctx, s.cfg.DBPath, templates, telemetry.ReadModelOptions{
-		ProviderLinks:   req.ProviderLinks,
-		TimeWindowHours: tw.Hours(),
-		TimeWindow:      tw,
-	})
-	core.Tracef("[read_model_perf] computeReadModel TOTAL: %dms (window=%s, accounts=%d, results=%d)",
-		time.Since(start).Milliseconds(), tw, len(req.Accounts), len(result))
-	return result, err
-}
-
-func (s *Service) refreshReadModelCacheAsync(
-	parent context.Context,
-	cacheKey string,
-	req ReadModelRequest,
-	timeout time.Duration,
-) {
-	if !s.rmCache.beginRefresh(cacheKey) {
-		return
-	}
-	go func() {
-		defer s.rmCache.endRefresh(cacheKey)
-		refreshCtx, cancel := context.WithTimeout(parent, timeout)
-		defer cancel()
-		snapshots, err := s.computeReadModel(refreshCtx, req)
-		if err != nil {
-			if s.shouldLog("read_model_cache_refresh_error", 8*time.Second) {
-				s.warnf("read_model_cache_refresh_error", "error=%v", err)
-			}
-			return
-		}
-		s.rmCache.set(cacheKey, snapshots)
-	}()
-}
-
-func (s *Service) backgroundContext() context.Context {
-	if s != nil && s.ctx != nil {
-		return s.ctx
-	}
-	return context.Background()
-}
-
-func (s *Service) runReadModelCacheLoop(ctx context.Context) {
-	if s == nil {
-		return
-	}
-
-	interval := s.cfg.PollInterval / 2
-	interval = max(5*time.Second, min(30*time.Second, interval))
-
-	s.infof("read_model_cache_loop_start", "interval=%s", interval)
-	s.refreshReadModelCacheFromConfig(ctx)
-
-	ticker := time.NewTicker(interval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			s.infof("read_model_cache_loop_stop", "reason=context_done")
-			return
-		case <-ticker.C:
-			s.refreshReadModelCacheFromConfig(ctx)
-		}
-	}
-}
-
-func (s *Service) refreshReadModelCacheFromConfig(ctx context.Context) {
-	req, err := BuildReadModelRequestFromConfig()
-	if err != nil {
-		if s.shouldLog("read_model_cache_config_error", 15*time.Second) {
-			s.warnf("read_model_cache_config_error", "error=%v", err)
-		}
-		return
-	}
-	if len(req.Accounts) == 0 {
-		return
-	}
-	cacheKey := ReadModelRequestKey(req)
-	s.refreshReadModelCacheAsync(ctx, cacheKey, req, 60*time.Second)
 }
 
 // --- Collection loop ---
@@ -961,157 +836,6 @@ func EnsureSocketPathAvailable(socketPath string) error {
 	return nil
 }
 
-func (s *Service) handleHealth(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, map[string]any{
-		"status":                 "ok",
-		"daemon_version":         strings.TrimSpace(version.Version),
-		"api_version":            APIVersion,
-		"integration_version":    integrations.IntegrationVersion,
-		"provider_registry_hash": ProviderRegistryHash(),
-	})
-}
-
-func (s *Service) handleHook(w http.ResponseWriter, r *http.Request) {
-	started := time.Now()
-	if r.Method != http.MethodPost {
-		writeJSONError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-
-	sourceName := strings.TrimPrefix(strings.TrimSpace(r.URL.Path), "/v1/hook/")
-	sourceName = strings.TrimSpace(strings.Trim(sourceName, "/"))
-	if sourceName == "" {
-		writeJSONError(w, http.StatusBadRequest, "missing hook source")
-		return
-	}
-	source, ok := providers.TelemetrySourceBySystem(sourceName)
-	if !ok {
-		writeJSONError(w, http.StatusBadRequest, fmt.Sprintf("unknown hook source %q", sourceName))
-		return
-	}
-
-	payload, err := io.ReadAll(r.Body)
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, "read payload failed")
-		return
-	}
-	if len(strings.TrimSpace(string(payload))) == 0 {
-		writeJSONError(w, http.StatusBadRequest, "empty payload")
-		return
-	}
-
-	accountID := strings.TrimSpace(r.URL.Query().Get("account_id"))
-	reqs, err := telemetry.ParseSourceHookPayload(source, payload, source.DefaultCollectOptions(), accountID)
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, fmt.Sprintf("parse hook payload: %v", err))
-		return
-	}
-	if len(reqs) == 0 {
-		writeJSON(w, http.StatusOK, HookResponse{Source: sourceName})
-		return
-	}
-
-	tally, _ := s.ingestBatch(r.Context(), reqs)
-	var warnings []string
-	if tally.failed > 0 {
-		warnings = append(warnings, fmt.Sprintf("%d ingest failures", tally.failed))
-	}
-
-	writeJSON(w, http.StatusOK, HookResponse{
-		Source:    sourceName,
-		Enqueued:  len(reqs),
-		Processed: tally.processed,
-		Ingested:  tally.ingested,
-		Deduped:   tally.deduped,
-		Failed:    tally.failed,
-		Warnings:  warnings,
-	})
-
-	durationMs := time.Since(started).Milliseconds()
-	logLevel := "hook_ingest"
-	shouldLog := tally.failed > 0 || s.shouldLog("hook_ingest_"+sourceName, 3*time.Second)
-	if !shouldLog {
-		return
-	}
-	if tally.failed > 0 {
-		s.warnf(logLevel,
-			"source=%s account_id=%q duration_ms=%d enqueued=%d processed=%d ingested=%d deduped=%d failed=%d",
-			sourceName, accountID, durationMs,
-			len(reqs), tally.processed, tally.ingested, tally.deduped, tally.failed,
-		)
-	} else {
-		s.infof(logLevel,
-			"source=%s account_id=%q duration_ms=%d enqueued=%d processed=%d ingested=%d deduped=%d failed=%d",
-			sourceName, accountID, durationMs,
-			len(reqs), tally.processed, tally.ingested, tally.deduped, tally.failed,
-		)
-	}
-}
-
-func (s *Service) handleReadModel(w http.ResponseWriter, r *http.Request) {
-	started := time.Now()
-	if r.Method != http.MethodPost {
-		writeJSONError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-
-	var req ReadModelRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeJSONError(w, http.StatusBadRequest, fmt.Sprintf("decode read-model request: %v", err))
-		return
-	}
-
-	if len(req.Accounts) == 0 {
-		configReq, configErr := BuildReadModelRequestFromConfig()
-		if configErr != nil || len(configReq.Accounts) == 0 {
-			writeJSON(w, http.StatusOK, ReadModelResponse{Snapshots: map[string]core.UsageSnapshot{}})
-			return
-		}
-		req = configReq
-	}
-
-	cacheKey := ReadModelRequestKey(req)
-	if cached, cachedAt, ok := s.rmCache.get(cacheKey); ok {
-		core.Tracef("[read_model] cache hit key=%s age=%s providers=%d", cacheKey, time.Since(cachedAt).Round(time.Millisecond), len(cached))
-		for id, snap := range cached {
-			core.Tracef("[read_model]   %s: %d metrics", id, len(snap.Metrics))
-		}
-		writeJSON(w, http.StatusOK, ReadModelResponse{Snapshots: cached})
-		if time.Since(cachedAt) > 2*time.Second {
-			s.refreshReadModelCacheAsync(s.backgroundContext(), cacheKey, req, 60*time.Second)
-		}
-		return
-	}
-
-	computeCtx, cancel := context.WithTimeout(r.Context(), 5*time.Second)
-	snapshots, err := s.computeReadModel(computeCtx, req)
-	cancel()
-	if err == nil && len(snapshots) > 0 {
-		s.rmCache.set(cacheKey, snapshots)
-		writeJSON(w, http.StatusOK, ReadModelResponse{Snapshots: snapshots})
-		return
-	}
-
-	if err != nil && s.shouldLog("read_model_cache_miss_compute_error", 8*time.Second) {
-		s.warnf("read_model_cache_miss_compute_error", "error=%v", err)
-	}
-
-	s.refreshReadModelCacheAsync(s.backgroundContext(), cacheKey, req, 60*time.Second)
-	snapshots = ReadModelTemplatesFromRequest(req, DisabledAccountsFromConfig())
-	writeJSON(w, http.StatusOK, ReadModelResponse{Snapshots: snapshots})
-	durationMs := time.Since(started).Milliseconds()
-	if durationMs >= 1200 && s.shouldLog("read_model_slow", 30*time.Second) {
-		s.infof(
-			"read_model_slow",
-			"duration_ms=%d requested_accounts=%d returned_snapshots=%d provider_links=%d",
-			durationMs,
-			len(req.Accounts),
-			len(snapshots),
-			len(req.ProviderLinks),
-		)
-	}
-}
-
 // --- Helpers ---
 
 func buildCollectors() []telemetry.Collector {
@@ -1171,14 +895,4 @@ func FlushInBatches(ctx context.Context, pipeline *telemetry.Pipeline, maxTotal 
 	}
 
 	return accum, warnings
-}
-
-func writeJSON(w http.ResponseWriter, status int, payload any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	_ = json.NewEncoder(w).Encode(payload)
-}
-
-func writeJSONError(w http.ResponseWriter, status int, message string) {
-	writeJSON(w, status, map[string]string{"error": message})
 }
